Update 2.1
It has been a while since I updated this branch! Been busy with certifications and volunteering! This update introduced new bug fixes on the program, alongside the use of youtube-dl to do the work of filtering the youtube playlist instead of BeautifulSoup. Might have been due to change of Youtube data structuring